feat: add analytics-client
This commit introduces a new client for fetching trade analytics data.  It uses gRPC for communication and includes a command-line interface for easy interaction.  The client connects to an analytics server (currently assumed to be running locally on port 50051) and retrieves trade analytics for a specified symbol within a given time range.  The results are then displayed to the user.